pyproject.toml
[tool.poetry]
name = "repl_python3_RPG-Modules-and-Maps-Hand-in-Version"
version = "0.1.0"
description = ""
authors = ["humzahmalik07 <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
action = "^1.4.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
